fix test, remove unnecessary hashCode and equals methods in sub classes
also start moving id only to a common super classes
In the future, we try to have one common super class with id
