+++ src/main/java/de/tum/in/www1/artemis/domain/FileUploadExercise.java
@@ -1,7 +1,6 @@
 package de.tum.in.www1.artemis.domain;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 import javax.persistence.Column;
 import javax.persistence.DiscriminatorValue;
@@ -11,7 +10,6 @@ import javax.persistence.Lob;
 /**
  * A FileUploadExercise.
  */
-
 @Entity
 @DiscriminatorValue(value = "F")
 public class FileUploadExercise extends Exercise implements Serializable {
@@ -22,8 +20,6 @@ public class FileUploadExercise extends Exercise implements Serializable {
 
     @Column(name = "filePattern")
     private String filePattern;
-
-    // jhipster-needle-entity-add-field - Jhipster will add fields here, do not remove
 
     public String getFilePattern() {
         return filePattern;
@@ -49,26 +45,6 @@ public class FileUploadExercise extends Exercise implements Serializable {
 
     public void setSampleSolution(String sampleSolution) {
         this.sampleSolution = sampleSolution;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        FileUploadExercise fileUploadExercise = (FileUploadExercise) o;
-        if (fileUploadExercise.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), fileUploadExercise.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
