+++ src/main/java/de/tum/in/www1/artemis/domain/ProgrammingSubmission.java
@@ -3,7 +3,6 @@ package de.tum.in.www1.artemis.domain;
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 
 import javax.persistence.*;
 
@@ -37,8 +36,6 @@ public class ProgrammingSubmission extends Submission implements Serializable {
     @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
     private List<BuildLogEntry> buildLogEntries = new ArrayList<>();
 
-    // jhipster-needle-entity-add-field - JHipster will add fields here, do not remove
-
     public String getCommitHash() {
         return commitHash;
     }
@@ -51,7 +48,6 @@ public class ProgrammingSubmission extends Submission implements Serializable {
     public void setCommitHash(String commitHash) {
         this.commitHash = commitHash;
     }
-    // jhipster-needle-entity-add-getters-setters - JHipster will add getters and setters here, do not remove
 
     public boolean isBuildFailed() {
         return buildFailed;
@@ -75,23 +71,6 @@ public class ProgrammingSubmission extends Submission implements Serializable {
 
     public void setBuildLogEntries(List<BuildLogEntry> buildLogEntries) {
         this.buildLogEntries = buildLogEntries;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (!(o instanceof ProgrammingSubmission))
-            return false;
-        if (!super.equals(o))
-            return false;
-        ProgrammingSubmission that = (ProgrammingSubmission) o;
-        return buildFailed == that.buildFailed && buildArtifact == that.buildArtifact && Objects.equals(commitHash, that.commitHash);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
