+++ src/main/java/de/tum/in/www1/artemis/domain/TextExercise.java
@@ -2,7 +2,6 @@ package de.tum.in.www1.artemis.domain;
 
 import java.io.Serializable;
 import java.util.List;
-import java.util.Objects;
 
 import javax.persistence.*;
 
@@ -48,26 +47,6 @@ public class TextExercise extends Exercise implements Serializable {
     public void filterSensitiveInformation() {
         setSampleSolution(null);
         super.filterSensitiveInformation();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        TextExercise textExercise = (TextExercise) o;
-        if (textExercise.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), textExercise.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
