+++ src/main/java/de/tum/in/www1/artemis/domain/exam/ExamSession.java
@@ -1,8 +1,5 @@
 package de.tum.in.www1.artemis.domain.exam;
 
-import java.io.Serializable;
-import java.util.Objects;
-
 import javax.persistence.*;
 
 import org.hibernate.annotations.Cache;
@@ -18,12 +15,7 @@ import inet.ipaddr.IPAddressString;
 @Table(name = "exam_session")
 @Cache(usage = CacheConcurrencyStrategy.NONSTRICT_READ_WRITE)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class ExamSession extends AbstractAuditingEntity implements Serializable {
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+public class ExamSession extends AbstractAuditingEntity {
 
     @ManyToOne
     @JoinColumn(name = "student_exam_id")
@@ -45,14 +37,6 @@ public class ExamSession extends AbstractAuditingEntity implements Serializable 
     private String ipAddress;
 
     public ExamSession() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public StudentExam getStudentExam() {
@@ -110,20 +94,4 @@ public class ExamSession extends AbstractAuditingEntity implements Serializable 
         setInstanceId(null);
         setIpAddress(null);
     }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (o == null || getClass() != o.getClass())
-            return false;
-        ExamSession that = (ExamSession) o;
-        return Objects.equals(getId(), that.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
-    }
-
 }
