+++ src/main/java/de/tum/in/www1/artemis/domain/modeling/ModelingExercise.java
@@ -1,7 +1,6 @@
 package de.tum.in.www1.artemis.domain.modeling;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 import javax.persistence.*;
 
@@ -28,8 +27,6 @@ public class ModelingExercise extends Exercise implements Serializable {
     @Column(name = "sample_solution_explanation")
     @Lob
     private String sampleSolutionExplanation;
-
-    // jhipster-needle-entity-add-field - Jhipster will add fields here, do not remove
 
     public DiagramType getDiagramType() {
         return diagramType;
@@ -70,8 +67,6 @@ public class ModelingExercise extends Exercise implements Serializable {
         this.sampleSolutionExplanation = sampleSolutionExplanation;
     }
 
-    // jhipster-needle-entity-add-getters-setters - Jhipster will add getters and setters here, do not remove
-
     /**
      * set all sensitive information to null, so no info with respect to the solution gets leaked to students through json
      */
@@ -80,26 +75,6 @@ public class ModelingExercise extends Exercise implements Serializable {
         setSampleSolutionModel(null);
         setSampleSolutionExplanation(null);
         super.filterSensitiveInformation();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        ModelingExercise modelingExercise = (ModelingExercise) o;
-        if (modelingExercise.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), modelingExercise.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
