+++ src/main/java/de/tum/in/www1/artemis/domain/modeling/ModelingSubmission.java
@@ -1,7 +1,6 @@
 package de.tum.in.www1.artemis.domain.modeling;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 import javax.persistence.Column;
 import javax.persistence.DiscriminatorValue;
@@ -55,28 +54,6 @@ public class ModelingSubmission extends Submission implements Serializable {
 
     public void setExplanationText(String explanationText) {
         this.explanationText = explanationText;
-    }
-
-    // jhipster-needle-entity-add-getters-setters - JHipster will add getters and setters here, do not remove
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        ModelingSubmission modelingSubmission = (ModelingSubmission) o;
-        if (modelingSubmission.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), modelingSubmission.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
