+++ src/main/java/de/tum/in/www1/artemis/domain/notification/SingleUserNotification.java
@@ -2,7 +2,6 @@ package de.tum.in.www1.artemis.domain.notification;
 
 import java.io.Serializable;
 import java.time.ZonedDateTime;
-import java.util.Objects;
 
 import javax.persistence.DiscriminatorValue;
 import javax.persistence.Entity;
@@ -25,7 +24,6 @@ public class SingleUserNotification extends Notification implements Serializable
     @ManyToOne
     private User recipient;
 
-    // jhipster-needle-entity-add-field - JHipster will add fields here, do not remove
     public User getRecipient() {
         return recipient;
     }
@@ -38,7 +36,6 @@ public class SingleUserNotification extends Notification implements Serializable
     public void setRecipient(User user) {
         this.recipient = user;
     }
-    // jhipster-needle-entity-add-getters-setters - JHipster will add getters and setters here, do not remove
 
     public String getTopic() {
         return "/topic/user/" + getRecipient().getId() + "/notifications";
@@ -85,26 +82,6 @@ public class SingleUserNotification extends Notification implements Serializable
         target.addProperty("course", studentQuestionAnswer.getQuestion().getLecture().getCourse().getId());
         target.addProperty("mainPage", "courses");
         return target.toString();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        SingleUserNotification singleUserNotification = (SingleUserNotification) o;
-        if (singleUserNotification.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), singleUserNotification.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
