+++ src/main/java/de/tum/in/www1/artemis/domain/notification/SystemNotification.java
@@ -2,7 +2,6 @@ package de.tum.in.www1.artemis.domain.notification;
 
 import java.io.Serializable;
 import java.time.ZonedDateTime;
-import java.util.Objects;
 
 import javax.persistence.*;
 
@@ -24,7 +23,6 @@ public class SystemNotification extends Notification implements Serializable {
     @Column(name = "jhi_type")
     private SystemNotificationType type;
 
-    // jhipster-needle-entity-add-field - JHipster will add fields here, do not remove
     public ZonedDateTime getExpireDate() {
         return expireDate;
     }
@@ -49,27 +47,6 @@ public class SystemNotification extends Notification implements Serializable {
 
     public void setType(SystemNotificationType type) {
         this.type = type;
-    }
-    // jhipster-needle-entity-add-getters-setters - JHipster will add getters and setters here, do not remove
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        SystemNotification systemNotification = (SystemNotification) o;
-        if (systemNotification.getId() == null || getId() == null) {
-            return false;
-        }
-        return Objects.equals(getId(), systemNotification.getId());
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(getId());
     }
 
     @Override
